Avoid overwriting model scopes in subclass
jsonb_accessor overwrites the scopes in each subclass. This is not
required as it is always the same method and causes a warning
in active record logging:

  WARN -- : Creating scope :options_where. Overwriting existing method
  Model.options_where.

This adds a logger to the rspec tests and will treat warnings
in rspec tests as error.
It also changes jsonb_accessor to not create the scopes on subclasses.