Fix performance issues with how jsonb_accessor clears dirty tracking
Closes #120

Previously, the jsonb_accessor after_init callback
would clear ALL of the models' attributes, and not
just the jsonb_accessor attributes. This had the
effect of forcefully deserializing and then
immediately reserializing ALL of the models'
attributes, which was causing a massive performance
hit on our app on a model that uses the `serialize`
API on another attribute.

This change makes sure to only reset the dirty-tracking
on the jsonb_accessor attributes.